% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alertness.R
\docType{data}
\name{alertness}
\alias{alertness}
\title{alertness}
\format{
A data frame with 720 rows and 6 variables:
\describe{
  \item{PartID}{Participant ID}
  \item{Sex}{Between subjects variable representing biological sex}
  \item{Group}{Between subjects variable}
  \item{Condition}{Within subjects variable}
  \item{Time}{Within subjects variable}
  \item{Alertness}{Dependent variable indicating the participants level of alterness}
}
}
\usage{
alertness
}
\description{
A dataset of how alertness changes in response to 2 different experimental conditions across 3 time points for each of two groups.
}
\keyword{datasets}
