% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antitibbler.R
\name{antitibbler}
\alias{antitibbler}
\title{antitibbler}
\usage{
antitibbler(datain)
}
\arguments{
\item{datain}{tibble of data}
}
\description{
Function to return tibble to data frame.
}
\author{
Matthew B. Pontifex, \email{pontifex@msu.edu}, September 13, 2021
}
