% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BootstrapRmimicLMcontrast.R
\name{BootstrapRmimicLMcontrast}
\alias{BootstrapRmimicLMcontrast}
\title{BootstrapRmimicLMcontrast}
\usage{
BootstrapRmimicLMcontrast(
  basefit,
  fit,
  repetitions = 999,
  resample_min = FALSE,
  resample_max = FALSE,
  studywiseAlpha = 0.05,
  confidenceinterval = 0.95
)
}
\arguments{
\item{fit}{fit from a basic linear model}

\item{repetitions}{number of bootstrap repetitions. Default is 999.}

\item{resample_min}{minimum number of samples to include when resampling.}

\item{resample_max}{minimum number of samples to include when resampling.}

\item{studywiseAlpha}{Decimal representation of alpha level. Default 0.05.}

\item{confidenceinterval}{Decimal representation of confidence interval. Default 0.95.}

\item{altfit}{fit from a linear model of interest}
}
\value{
\item{stats}{ANOVA summary table.}
\item{modelcheck}{chisquared summary table.}
\item{changestats}{ANOVA summary table of change statistics.}
\item{coefficients}{A summary table of regression coefficients.}
}
\description{
Compute results for bootstrapped regression analysis with effect size and confidence intervals. This function takes stats::lm fits for a base model and the model of interest and calculates statistics for the model of interest relative to the base model. For logistic regression, pseudo r2 values are reported using Tjurs 2009 approach.
}
\examples{

    # Compute Model from mtcars dataset
    basefit <- lm(mpg ~ am + wt, data = mtcars)
    fit <- lm(mpg ~ am + wt + qsec, data = mtcars)
    result <- BootstrapRmimicLMcontrast(basefit, fit,
         repetitions=99, resample_min=32, resample_max=96)

}
\author{
Matthew B. Pontifex, \email{pontifex@msu.edu}, March 25, 2025
}
