% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterthreshold1d.R
\name{clusterthreshold1d}
\alias{clusterthreshold1d}
\title{clusterthreshold1d}
\usage{
clusterthreshold1d(datain, crit = 0.05,
    clustersize = 5, direction = 'LessThan')
}
\arguments{
\item{datain}{Data vector.}

\item{crit}{Decimal representation of critical value. Default 0.05.}

\item{clustersize}{The cluster size of values meeting critial value.}

\item{direction}{Directional parameter for application of the critical value. Default is LessThan.}
}
\value{
A numeric logical vector corresponding to those points that meet the criteria.
}
\description{
Calculate contiguous clusters of locations in a 1D array that are above or below some threshold and with some minimum size.
}
\examples{
    # Returns a vector corresponding to those values below 0.05 in 5 sample clusters
    result <- clusterthreshold1d(stats::runif(100), crit = 0.05,
        clustersize = 5, direction = 'LessThan')

}
\author{
Matthew B. Pontifex, \email{pontifex@msu.edu}, June 24, 2019
}
