% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorpalette.R
\name{colorpalette}
\alias{colorpalette}
\title{colorpalette}
\usage{
colorpalette()
}
\description{
Function to return a set of standard color pallets
}
\examples{
    # Round a randomly generated number between 0 and 1.
    workingpallete <- Rmimic::colorpalette()

}
\author{
Matthew B. Pontifex, \email{pontifex@msu.edu}, June 2, 2025
}
