% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computechange.R
\name{computechange}
\alias{computechange}
\title{computechange}
\usage{
computechange(
  data,
  method = NULL,
  dependentvariable = NULL,
  subjectid = NULL,
  within = NULL,
  relativeto = NULL,
  control = NULL
)
}
\arguments{
\item{data}{Database containing data}

\item{method}{Parameter to select if the computation should be done as Change or PercentChange. Default is Change.}

\item{dependentvariable}{Dependent Variable label}

\item{subjectid}{Subject ID label.}

\item{within}{Within subjects factor labels to focus on.}

\item{relativeto}{Factor level to use as the baseline. Default is to use the lowest factor level.}

\item{control}{Column labels for factors to account for.}
}
\value{
\item{data}{Database with an additional variable reflecitng the change in the dependent variable.}
}
\description{
Function to compute the change or difference between conditions.
}
\examples{
    mockdatabase <- data.frame("ID" = rep_len(1:20,60), 
    "Time" = c(rep_len("Time1",20),rep_len("Time2",20),rep_len("Time3",20)), "X" = runif(60))
    mockdatabase <- computechange(mockdatabase, dependentvariable='X', subjectid='ID', within='Time')

}
\author{
Matthew B. Pontifex, \email{pontifex@msu.edu}, November 20, 2019
}
