% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{correlation}
\alias{correlation}
\title{correlation}
\usage{
correlation(
  variables = FALSE,
  partial = FALSE,
  data = FALSE,
  method = FALSE,
  listwise = TRUE,
  studywiseAlpha = 0.05,
  confidenceinterval = 0.95,
  verbose = TRUE
)
}
\arguments{
\item{variables}{Variable name or list of variables to compute descriptives for.}

\item{partial}{Parameter to specify the variable name to use for partial correlations.}

\item{data}{Data frame containing the variables of interest.}

\item{method}{Specifies the type of correlation. Options are pearson (default), spearman, or kendall.}

\item{listwise}{Boolean operator for if the correlations should be computed using list-wise deletion for missing values.}

\item{studywiseAlpha}{Decimal representation of alpha level. Default 0.05.}

\item{confidenceinterval}{Decimal representation of confidence interval. Default 0.95.}

\item{verbose}{Parameter to print all output to console. Default is TRUE.}
}
\value{
A list with the following matrix elements:
\item{Comparison}{Comparisons made.}
\item{n}{Number of subjects.}
\item{estimate}{Correlation between comparisons.}
\item{\samp{p.value}}{Correlation p value.}
\item{statistic}{The value of the test statistic.}
\item{parameter}{The degrees of freedom of the test statistic.}
\item{\samp{conf.int.lower}}{Minimum confidence interval of correlation.}
\item{\samp{conf.int.upper}}{Maximum confidence interval of correlation.}
\item{method}{A character string indicating how the association was measured.}
\item{null.value}{The value of the association measure under the null hypothesis.}
\item{alternative}{A character string describing the alternative hypothesis.}
\item{textoutput}{A character string of the test result.}
}
\description{
Compute SPSS style correlation or partial correlation, with optional parameters for the approach.
}
\examples{
    # Compute correlation.
    working <- data.frame("X" = runif(100), "Y" = runif(100), "Z" = runif(100))
    sR <- correlation(variables = c('X','Y','Z'), data = working,
        method="pearson", listwise=TRUE)

}
\author{
Matthew B. Pontifex, \email{pontifex@msu.edu}, May 4, 2020
}
