% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframenamecheck.R
\name{dataframenamecheck}
\alias{dataframenamecheck}
\title{dataframenamecheck}
\usage{
dataframenamecheck(datain)
}
\arguments{
\item{datain}{dataframe}
}
\description{
Function to make sure dataframe names do not create an issue.
}
\author{
Matthew B. Pontifex, \email{pontifex@msu.edu}, July 20, 2021
}
