% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decimalplaces.R
\name{decimalplaces}
\alias{decimalplaces}
\title{decimalplaces}
\usage{
decimalplaces(datavector)
}
\arguments{
\item{datavector}{Vector of data}
}
\value{
\item{precisionvector}{Integer of maximum number of decimal places in a vector.}
}
\description{
Obtain the number of decimal places of precision in a vector
}
\examples{
    precisioninteger <- decimalplaces(334.3410000000000000)

}
