% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{descriptives}
\alias{descriptives}
\title{descriptives}
\usage{
descriptives(comparisondataframe, groupvariable = FALSE,
  verbose = TRUE, verbosedescriptives = TRUE,
  verbosefrequencies = TRUE)
}
\arguments{
\item{comparisondataframe}{Data frame with each column containing a variable (includes both data and grouping variables if group descriptives are specified)}

\item{groupvariable}{Variable name or list of variables to use for grouping.}

\item{verbose}{Parameter to print all output to console. Default is TRUE.}

\item{verbosedescriptives}{Parameter to print descriptives output to console. Default is TRUE.}

\item{verbosefrequencies}{Parameter to print frequency output to console. Default is TRUE.}
}
\value{
\item{datamatrix}{Data table of descriptive statistics.}
}
\description{
Compute SPSS style descriptives.
}
\examples{

    # Compute descriptive statistics
    desc <- descriptives(mtcars, groupvariable=FALSE, 
    verbose=FALSE, verbosedescriptives=FALSE,
    verbosefrequencies=FALSE)

    # Compute descriptive statistics by group
    desc <- descriptives(mtcars, groupvariable=c("cyl"), 
    verbose=TRUE, verbosedescriptives=TRUE, 
    verbosefrequencies=FALSE)

}
\author{
Matthew B. Pontifex, \email{pontifex@msu.edu}, October 3, 2019
}
