% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determineallpossiblecombinations.R
\name{determineallpossiblecombinations}
\alias{determineallpossiblecombinations}
\title{determineallpossiblecombinations}
\usage{
determineallpossiblecombinations(datain)
}
\arguments{
\item{datain}{List of factors.}
}
\value{
A list of all possible combinations.
}
\description{
Function to determine all possible factor combinations.
}
\examples{
    # Determine all combinations of a generic list
    outlist <- determineallpossiblecombinations(c('A', 'B', 'C'))

}
\author{
Matthew B. Pontifex, \email{pontifex@msu.edu}, November 13, 2017
}
