% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixname.r
\name{fixname}
\alias{fixname}
\title{fixname}
\usage{
fixname(tmpname)
}
\arguments{
\item{tmpname}{names}
}
\description{
Function to check variable names for issues and fix them.
}
\author{
Matthew B. Pontifex, \email{pontifex@msu.edu}, September 13, 2021
}
