% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradwakefulness.R
\docType{data}
\name{gradwakefulness}
\alias{gradwakefulness}
\title{gradwakefulness}
\format{
A data frame with 200 rows and 3 variables:
\describe{
  \item{HoursAwake}{Number of hours awake.}
  \item{CaffeineConsumption}{Milligrams of Caffeine consumed.}
  \item{Wakefulness}{Rating of how alert the individual feels.}
}
}
\usage{
gradwakefulness
}
\description{
A dataset of how 3 different drugs and 3 different dosages of those drugs impact an individuals level of alterness. Participants  .
}
\keyword{datasets}
