% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyoutliers.R
\name{identifyoutliers}
\alias{identifyoutliers}
\title{identifyoutliers}
\usage{
identifyoutliers(
  datavector,
  iqrlimit = 3,
  direction = "NA",
  minuniquesamples = 10,
  verbose = TRUE
)
}
\arguments{
\item{datavector}{Vector of data}

\item{iqrlimit}{The interquartile limit. Default is 3 consistent with SPSS' identification of extreme outliers.}

\item{direction}{Parameter to restrict identification of outliers to a particular tail. Default is two tailed. Parameters are: loweronly or upperonly.}

\item{minuniquesamples}{Minimum number of samples for outlier detection. Default is 10.}

\item{verbose}{Parameter to print the identified outliers.}
}
\value{
\item{datavector}{Vector with outliers replaced with NA.}
}
\description{
Function to identify outliers based upon the interquartile range (as SPSS does for boxplots) and replace those values with NA. SPSS identifies mild outliers (1.5 x IQR) with circles and extreme outliers (3 x IQR) with asterisks.
}
\examples{
    # Populate generic data frame
    testvect <- runif(100,1,10)

    # Insert outlying data points
    testvect[6] <- -1000
    testvect[10] <- 1000

    testvect <- identifyoutliers(testvect, iqrlimit = 3, verbose=TRUE)

}
\author{
Matthew B. Pontifex, \email{pontifex@msu.edu}, September 12, 2019
}
