% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyoutliersdataframe.R
\name{identifyoutliersdataframe}
\alias{identifyoutliersdataframe}
\title{identifyoutliersdataframe}
\usage{
identifyoutliersdataframe(
  data,
  variables = FALSE,
  iqrlimit = 3,
  direction = "NA",
  minuniquesamples = 10,
  verbose = TRUE
)
}
\arguments{
\item{data}{Data frame containing the variables of interest.}

\item{variables}{Variable name or list of variables to impute.}

\item{iqrlimit}{The interquartile limit. Default is 3 consistent with SPSS' identification of extreme outliers.}

\item{direction}{Parameter to restrict identification of outliers to a particular tail. Default is two tailed. Parameters are: loweronly or upperonly.}

\item{minuniquesamples}{Minimum number of samples for outlier detection. Default is 10.}

\item{verbose}{Parameter to print the identified outliers.}
}
\value{
\item{datta}{Data frame with outliers replaced with NA.}
}
\description{
Function to identify outliers in a data frame based upon the interquartile range (as SPSS does for boxplots) and replace those values with NA. SPSS identifies mild outliers (1.5 x IQR) with circles and extreme outliers (3 x IQR) with asterisks.
}
\author{
Matthew B. Pontifex, \email{pontifex@msu.edu}, May 18, 2020
}
