% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmer2text.R
\name{lmer2text}
\alias{lmer2text}
\title{lmer2text}
\usage{
lmer2text(
  fit,
  model = NULL,
  df = NULL,
  numparticipants = NULL,
  numfactors = NULL,
  confidenceinterval = 0.95
)
}
\arguments{
\item{fit}{lmerTest model fit}

\item{model}{Parameter to select model type. Currently only ANOVA is available.}

\item{df}{Parameter to indicate what degrees of freedom approximation should be used. Default is Kenward-Roger. Other option is Shattertwaite.}

\item{numparticipants}{Number of unique participants in model.}

\item{numfactors}{Number of fixed and random factors in model.}

\item{confidenceinterval}{Decimal representation of confidence interval. Default 0.95.}
}
\value{
A list with:
\item{fit}{lmerTest object}
\item{ANOVA}{ANOVA table for fixed effects}
\item{RandomEffectsANOVA}{ANOVA table for random effects}
\item{Rsquared}{model R squared values}
}
\description{
Output lmerTest results in APA style format with effect sizes and confidence intervals.
}
\examples{

#Note, use lmerTest::lmer instead of lme4::lmer for inclusion of p values
result <- lmer2text(fit, df="Kenward-Roger", numparticipants=30, numfactors=4)

}
\author{
Matthew B. Pontifex, \email{pontifex@msu.edu}, December 19, 2019
}
