% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmerEffects.R
\name{lmerEffects}
\alias{lmerEffects}
\title{lmerEffects}
\usage{
lmerEffects(
  fit,
  dependentvariable = NULL,
  subjectid = NULL,
  within = NULL,
  df = NULL,
  confidenceinterval = 0.95,
  studywiseAlpha = 0.05,
  suppresstext = FALSE,
  smp = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{fit}{modelfit from the lmer function}

\item{dependentvariable}{text specifying the dependent variable label}

\item{subjectid}{text specifying the subject id label. If left NULL assumes all factors are between subjects}

\item{within}{text list specifying any variables that should be considered as within subjects}

\item{df}{Parameter to indicate what degrees of freedom approximation should be used. Default is Kenward-Roger. Other options are Shattertwaite or Traditional.}

\item{confidenceinterval}{Decimal representation of confidence interval. Default 0.95.}

\item{studywiseAlpha}{Decimal representation of alpha level. Default 0.05.}

\item{suppresstext}{Boolean parameter to avoid obtaining html formatted text results.}
}
\value{
\item{stats}{ANOVA summary table.}
\item{randomstats}{ANOVA summary table for random effects.}
\item{rsquared}{summary table for r squared portions.}
}
\description{
Compute SPSS style univariate ANOVA with effect size and confidence intervals using a multi-level model from the lme4 function.
}
\examples{

    fit <- lmerTest::lmer(Alertness ~ Group*Time + (1 | PartID), data = Rmimic::alertness)
    results <- Rmimic::lmerEffects(fit, dependentvariable = "Alertness", subjectid = "PartID", df = "Kenward-Roger")

}
\author{
Matthew B. Pontifex, \email{pontifex@msu.edu}, May 1, 2025
}
