% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmerEffects2text.R
\name{lmerEffects2text}
\alias{lmerEffects2text}
\title{lmerEffects2text}
\usage{
lmerEffects2text(
  res,
  tag = "",
  subtag = "",
  testconfidence = FALSE,
  significanceconfidence = FALSE
)
}
\arguments{
\item{res}{list containing data frames for stats, randomstats, and rsquared values}

\item{tag}{string containing ID tag for the text output}
}
\value{
res list with an added column for the text output
}
\description{
Sub function to clean up and round values from Rmimic::lmerEffects.
}
\author{
Matthew B. Pontifex, \email{pontifex@msu.edu}, March 28, 2025
}
