% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmerEffects_simpledesc.R
\name{lmerEffects_simpledesc}
\alias{lmerEffects_simpledesc}
\title{lmerEffects_simpledesc}
\usage{
lmerEffects_simpledesc(results)
}
\arguments{
\item{results}{list output from the lmerEffects function containing the collection of results}
}
\value{
results list output from the lmerEffects function
}
\description{
Subfunction of Rmimic::lmerEffects. Creates descriptives table.
}
\author{
Matthew B. Pontifex, \email{pontifex@msu.edu}, May 1, 2025
}
