% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmerEffectsBootstrap.R
\name{lmerEffectsBootstrap}
\alias{lmerEffectsBootstrap}
\title{lmerEffectsBootstrap}
\usage{
lmerEffectsBootstrap(
  results,
  repetitions,
  resample_min = NULL,
  resample_max = NULL,
  subsample = 0.96,
  inflation = 1,
  method = "default",
  average = "median",
  reporteddata = "simulated",
  progressbar = TRUE
)
}
\arguments{
\item{results}{List containing output from lmerEffects}

\item{repetitions}{Numeric entry indicating the number of repetitions to perform}

\item{resample_min}{Numeric entry indicating the minimum number of samples to include (with replacement). Default keeps the total samples the same. Ignored for method parametric.}

\item{resample_max}{Numeric entry indicating the maximum number of samples to include (with replacement). Default keeps the total samples the same. Ignored for method parametric.}

\item{subsample}{Numeric entry 0 to 1 indicating what percentage of the data to use for informing the parametric simulation. Ignored for method resample.}

\item{inflation}{Numeric entry indicating how many times the original sample to simulate too. Ignored for method resample.}

\item{method}{String entry indicating the approach. Default simulates data by drawing random samples from the conditional distribution of the outcome variable given the estimated model parameters using the simulate function. This keeps the data around the original mean and standard deviation, thus as sample size increases the effect size will also increase. Resample performs data resampling with replacement from the existing dataset. Parametric simulates data from a multivariate normal distribution using the MASS::mvrnorm function. This keeps the effect size the same as it allows the group variation to increase with larger samples. Nonparametric simulates data from a multivariate nonnormal distribution using the mnonr::unonr function. This keeps the effect size the same as it allows the group variation to increase with larger samples.}

\item{average}{text parameter to indicate how the results should be collapsed. Default is using the median. Other options is mean.}

\item{reporteddata}{text parameter to indicate if the posthoc text reports should use actual data (actual) or should report the simulated or resampled data means.}

\item{progressbar}{Boolean parameter for if a progress bar should be shown indicating the status of the function. Note that multi way interactions may take a substantial period of time to fully decompose.}
}
\value{
\item{results}{List containing boostrapped output.}
}
\description{
Wrapper to perform bootstrap analysis on an lmer object processed through the Rmimic::lmerEffects function.
}
\examples{
\dontrun{
    fit <- lmerTest::lmer(Alertness ~ Group*Time + (1 | PartID), data = Rmimic::alertness)
    results <- Rmimic::lmerEffects(fit, dependentvariable = "Alertness", subjectid = "PartID", df = "Kenward-Roger")
    results <- Rmimic::lmerPosthoc(results, between=c('Group'), within=c('Time'),
               covariates=NULL, planned=c('Group'), posthoccorrection="False Discovery Rate Control", progressbar=TRUE)
    results <- Rmimic::lmerEffectsBootstrap(results, repetitions=999)
    }

}
\author{
Matthew B. Pontifex, \email{pontifex@msu.edu}, May 1, 2025
}
