% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmerEffectsBootstrapSimulationConsolidator.R
\name{lmerEffectsBootstrapSimulationConsolidator}
\alias{lmerEffectsBootstrapSimulationConsolidator}
\title{lmerEffectsBootstrapSimulationConsolidator}
\usage{
lmerEffectsBootstrapSimulationConsolidator(
  results,
  average = "median",
  reporteddata = "simulated"
)
}
\arguments{
\item{results}{List containing output from lmerEffects}

\item{average}{text parameter to indicate how the results should be collapsed. Default is using the median. Other options is mean.}

\item{reporteddata}{text parameter to indicate if the posthoc text reports should use actual data (actual) or should report the simulated or resampled data means.}
}
\value{
\item{results}{List containing boostrapped output.}
}
\description{
Wrapper to perform bootstrap analysis on an lmer object processed through the Rmimic::lmerEffects function. This portion handles the consolidation of the models and summarizes them into a singular model.
}
\examples{
\dontrun{
    fit <- lmerTest::lmer(Alertness ~ Group*Time + (1 | PartID), data = Rmimic::alertness)
    results <- Rmimic::lmerEffects(fit, dependentvariable = "Alertness", subjectid = "PartID", df = "Kenward-Roger")
    results <- Rmimic::lmerPosthoc(results, between=c('Group'), within=c('Time'),
               covariates=NULL, planned=c('Group'), posthoccorrection="False Discovery Rate Control", progressbar=TRUE)
    results <- Rmimic::lmerEffectsBootstrapSimulationCreation(results, repetitions=999, tmpdir='/tempdirectory/')
    results <- Rmimic::lmerEffectsBootstrapSimulationConsolidator(results, average='median', reporteddata='actual')
    }

}
\author{
Matthew B. Pontifex, \email{pontifex@msu.edu}, July 23, 2025
}
