% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmerEffectsSummarize.R
\name{lmerEffectsSummarize}
\alias{lmerEffectsSummarize}
\title{lmerEffectsSummarize}
\usage{
lmerEffectsSummarize(
  results,
  descriptives = TRUE,
  modelfit = TRUE,
  randomeffects = TRUE,
  fixedeffects = TRUE,
  style = NULL,
  hide = NULL,
  show = "html",
  outputfile = NULL,
  tag = ""
)
}
\arguments{
\item{results}{list output from the lmerEffects function}

\item{descriptives}{boolean parameter to control if descriptive statistics are shown before the model result}

\item{modelfit}{boolean parameter to control if model design and fit statistics are shown}

\item{randomeffects}{boolean parameter to control if random effects statistics are shown}

\item{fixedeffects}{boolean parameter to control if fixed effects statistics are shown}

\item{hide}{list parameter specifying any effects to suppress}

\item{show}{boolean parameter to control if any results are printed}

\item{outputfile}{text parameter containing a full file path and file name with html extension. Default is NULL which does not write a file.}
}
\description{
Summary report of univariate ANOVA with effect size and confidence intervals using a multi-level model from the lme4 function. Outputted HTML file has additional functionality to expand and collapse results for ease of reading.
}
\examples{

    fit <- lmerTest::lmer(Alertness ~ Group*Time + (1 | PartID), data = Rmimic::alertness)
    results <- Rmimic::lmerEffects(fit, dependentvariable = "Alertness", subjectid = "PartID", df = "Kenward-Roger")
    Rmimic::lmerEffectsSummarize(results, show=TRUE, outputfile=NULL)

}
\author{
Matthew B. Pontifex, \email{pontifex@msu.edu}, May 1, 2025
}
