% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmerEffectsSummarizesubprocess.R
\name{lmerEffectsSummarizesubprocess}
\alias{lmerEffectsSummarizesubprocess}
\title{lmerEffectsSummarizesubprocess}
\usage{
lmerEffectsSummarizesubprocess(results, mainpasstag = "", mainfactorlevels = 0)
}
\arguments{
\item{mainpasstag}{text containing a string to tag the results with}

\item{mainfactorlevels}{numeric indicating the number of test levels}

\item{res}{list containing data frames for stats, randomstats, and rsquared values}
}
\description{
A subfunction of lmerEffectsSummarize to place the lmer effects into html.
}
\author{
Matthew B. Pontifex, \email{pontifex@msu.edu}, April 29, 2025
}
