% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmerModelstats2text.R
\name{lmerModelstats2text}
\alias{lmerModelstats2text}
\title{lmerModelstats2text}
\usage{
lmerModelstats2text(
  res,
  confidenceinterval = 0.95,
  studywiseAlpha = 0.05,
  testconfidence = FALSE,
  significanceconfidence = FALSE
)
}
\arguments{
\item{res}{list containing output from Rmimic::lmerModelstats}
}
\value{
res list with an added elements for the text output
}
\description{
Sub function to create text interepretations from Rmimic::lmerModelstats
}
\author{
Matthew B. Pontifex, \email{pontifex@msu.edu}, June 3, 2025
}
