% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmerModelstatsBootstrap.R
\name{lmerModelstatsBootstrap}
\alias{lmerModelstatsBootstrap}
\title{lmerModelstatsBootstrap}
\usage{
lmerModelstatsBootstrap(
  results,
  repetitions,
  between,
  within = NULL,
  subjectid = NULL,
  resample_min = NULL,
  resample_max = NULL,
  subsample = 0.96,
  inflation = 1,
  method = "default",
  average = "median",
  progressbar = TRUE
)
}
\arguments{
\item{results}{List containing output from lmerModelstats}

\item{repetitions}{Numeric entry indicating the number of repetitions to perform}

\item{between}{text list specifying the between subjects variables and any between subjects covariates}

\item{within}{text list specifying any variables that should be considered as within subjects. If null assumes all variables are between subjects.}

\item{subjectid}{text specifying the subject id label.}

\item{resample_min}{Numeric entry indicating the minimum number of samples to include (with replacement). Default keeps the total samples the same. Ignored for method parametric.}

\item{resample_max}{Numeric entry indicating the maximum number of samples to include (with replacement). Default keeps the total samples the same. Ignored for method parametric.}

\item{subsample}{Numeric entry 0 to 1 indicating what percentage of the data to use for informing the parametric simulation. Ignored for method resample.}

\item{inflation}{Numeric entry indicating how many times the original sample to simulate too. Ignored for method resample.}

\item{method}{String entry indicating the approach. Default simulates data by drawing random samples from the conditional distribution of the outcome variable given the estimated model parameters using the simulate function. Resample performs data resampling with replacement from the existing dataset. Parametric simulates data from a multivariate normal distribution using the MASS::mvrnorm function. Nonparametric simulates data from a multivariate nonnormal distribution using the mnonr::unonr function.}

\item{average}{text parameter to indicate how the results should be collapsed. Default is using the median. Other options is mean.}

\item{progressbar}{Boolean parameter for if a progress bar should be shown indicating the status of the function. Note that multi way interactions may take a substantial period of time to fully decompose.}
}
\value{
\item{results}{List containing boostrapped output.}
}
\description{
Wrapper to perform bootstrap analysis on an lmer object processed through the Rmimic::lmerModelstats function.
}
\examples{
\dontrun{
    altfit <- lmerTest::lmer(Alertness ~ Time + (1 | PartID), data = Rmimic::alertness)
    fit <- lmerTest::lmer(Alertness ~ Group + Time + (1 | PartID), data = Rmimic::alertness)
    results <- Rmimic::lmerModelstats(fit, altfit)
    results <- Rmimic::lmerModelstatsBootstrap(results, repetitions=999, between=c('Group'),
                                              within=c('Time'), subjectid = "PartID",
                                              subsample=0.96, method='default', progressbar=TRUE)
    }

}
\author{
Matthew B. Pontifex, \email{pontifex@msu.edu}, June 4, 2025
}
