% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmerPosthocCorrection.R
\name{lmerPosthocCorrection}
\alias{lmerPosthocCorrection}
\title{lmerPosthocCorrection}
\usage{
lmerPosthocCorrection(
  results,
  method = NULL,
  studywiseAlpha = NULL,
  FDRC = 0.05
)
}
\arguments{
\item{results}{list containing data frames for stats, randomstats, and rsquared values}

\item{method}{text indicating the correction to apply. Options are Bonferroni, Sidak, Holm-Bonferroni, or False Discovery Rate Control (default)}

\item{studywiseAlpha}{decimal indicating the alpha level}

\item{FDRC}{decimal specifying the false discovery rate control. Default is 0.05}
}
\description{
Correction of posthoc tests. Subfunction of lmerPosthoc. Corrections are made in place and returned.
}
\examples{

    fit <- lmerTest::lmer(Alertness ~ Group*Time + (1 | PartID), data = Rmimic::alertness)
    results <- Rmimic::lmerEffects(fit, dependentvariable = "Alertness", subjectid = "PartID", df = "Kenward-Roger")
    results <- Rmimic::lmerPosthoc(results, between=c('Group'), within=c('Time'),
               covariates=NULL, planned=c('Group'), posthoccorrection="none", progressbar=TRUE)
    results <- Rmimic::lmerPosthocCorrection(results, method="False Discovery Rate Control", studywiseAlpha=0.05, FDRC=0.05)      
               
}
\author{
Matthew B. Pontifex, \email{pontifex@msu.edu}, May 1, 2025
}
