% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmerPosthocCorrectionsubprocess.R
\name{lmerPosthocCorrectionsubprocess}
\alias{lmerPosthocCorrectionsubprocess}
\title{lmerPosthocCorrectionsubprocess}
\usage{
lmerPosthocCorrectionsubprocess(results)
}
\arguments{
\item{res}{list containing data frames for stats, randomstats, and rsquared values}
}
\description{
Subprocess of lmerPosthoc by way of the subprocess lmerPosthocCorrection. Obtains data tables for all posthoc tests and their locations.
}
\author{
Matthew B. Pontifex, \email{pontifex@msu.edu}, March 28, 2025
}
