% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmerPosthocsubprocess.R
\name{lmerPosthocsubprocess}
\alias{lmerPosthocsubprocess}
\title{lmerPosthocsubprocess}
\usage{
lmerPosthocsubprocess(results, effectofinterest, ...)
}
\arguments{
\item{effectofinterest}{text specifying the effect label to decompose}

\item{fit}{modelfit from the lmer function}

\item{dependentvariable}{text specifying the dependent variable label}

\item{subjectid}{text specifying the subject id label.}

\item{within}{text list specifying any variables that should be considered as within subjects}

\item{between}{text list specifying the between subjects variables}

\item{covariates}{text list specifying the variables that should be considered as non interactive covariates.}

\item{planned}{text list specifying any effect to show the post-hoc comparisons even if they are not significant.}

\item{df}{Parameter to indicate what degrees of freedom approximation should be used. Default is Kenward-Roger. Other options are Shattertwaite or Traditional.}

\item{confidenceinterval}{Decimal representation of confidence interval. Default 0.95.}

\item{studywiseAlpha}{Decimal representation of alpha level. Default 0.05.}

\item{posthoclimit}{integer specifying how many factors in an interaction should be broken down. Default is 6 indicating posthoc results will be provided for up to a 6 way interaction.}

\item{progressbar}{boolean parameter for if a progress bar should be shown indicating the status of the function. Note that multi way interactions may take a substantial period of time to fully decompose.}
}
\description{
Subfunction of Rmimic::lmerPosthoc. Performs posthoc tests.
}
\author{
Matthew B. Pontifex, \email{pontifex@msu.edu}, April 22, 2025
}
