% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmerPosthocsubprocessContrasts.R
\name{lmerPosthocsubprocessContrasts}
\alias{lmerPosthocsubprocessContrasts}
\title{lmerPosthocsubprocessContrasts}
\usage{
lmerPosthocsubprocessContrasts(
  tempfit,
  emmeansdf,
  effectofinterest,
  currentfactor,
  factorsinvolved,
  otherfactorsinvolved,
  dependentvariable,
  decomptext,
  subjectid,
  confidenceinterval,
  studywiseAlpha,
  within,
  factortag
)
}
\description{
Subfunction of Rmimic::lmerPosthoc. Performs posthoc test contrasts.
}
\author{
Matthew B. Pontifex, \email{pontifex@msu.edu}, August 19, 2025
}
