% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmerRegressionSummarize.R
\name{lmerRegressionSummarize}
\alias{lmerRegressionSummarize}
\title{lmerRegressionSummarize}
\usage{
lmerRegressionSummarize(results, show = "html", outputfile = NULL, tag = "")
}
\arguments{
\item{results}{list output from the lmerModelstats function}

\item{show}{boolean parameter to control if any results are printed}

\item{outputfile}{text parameter containing a full file path and file name with html extension. Default is NULL which does not write a file.}

\item{tag}{text parameter to include additional text in html output file}
}
\description{
Summary report of linear model with effect size and confidence intervals using a multi-level model from the lme4 function.
}
\examples{

    altfit <- lmerTest::lmer(Alertness ~ 1 + (1 | PartID), data = Rmimic::alertness)
    fit <- lmerTest::lmer(Alertness ~ Group + (1 | PartID), data = Rmimic::alertness)
    results <- Rmimic::lmerModelstats(fit, altfit)
    Rmimic::lmerRegressionSummarize(results, show='html', outputfile=NULL)

}
\author{
Matthew B. Pontifex, \email{pontifex@msu.edu}, May 1, 2025
}
