% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmerSimulate_conditionaldistribution.R
\name{lmerSimulate_conditionaldistribution}
\alias{lmerSimulate_conditionaldistribution}
\title{lmerSimulate_conditionaldistribution}
\usage{
lmerSimulate_conditionaldistribution(
  fit,
  dependentvariable = NULL,
  subjectid = NULL,
  between = NULL,
  targetN = NULL,
  subsample = NULL
)
}
\arguments{
\item{fit}{modelfit from the lmer function}

\item{dependentvariable}{text specifying the dependent variable label}

\item{subjectid}{text specifying the subject id label.}

\item{between}{text list specifying the between subjects variables}

\item{targetN}{numeric value specifying the number of subjects in the simulated dataset.}
}
\value{
fit modelfit from the lmer function
}
\description{
Function to create a simulated dataset using simulate of the requested sample size.
}
\author{
Matthew B. Pontifex, \email{pontifex@msu.edu}, May 13, 2025
}
