% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediate2text.R
\name{mediate2text}
\alias{mediate2text}
\title{mediate2text}
\usage{
mediate2text(fit, studywiseAlpha = 0.05, verbose = TRUE)
}
\arguments{
\item{fit}{mediate model fit}

\item{studywiseAlpha}{Decimal representation of alpha level. Default 0.05.}

\item{verbose}{Boolean operator for if interpretations of the statistics should be printed. Default is TRUE.}
}
\value{
A list with:
\item{fit}{mediate object}
\item{summary}{summary object}
\item{paths}{table for path effects}
\item{interpretation}{text interpretation}
\item{propmediated}{proportion of the effect mediated}
}
\description{
Output mediate results in a more intelligible format.
}
\examples{


}
\author{
Matthew B. Pontifex, \email{pontifex@msu.edu}, October 1, 2020
}
