% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mimicR-package.R
\docType{package}
\name{mimicR-package}
\alias{mimicR}
\alias{mimicR-package}
\title{mimicR: R Interface to mimic C++ Functions}
\description{
Provides an R interface to the C++ peptide mimic functions.
}
\keyword{internal}
