% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motherchildBMI.R
\docType{data}
\name{motherchildBMI}
\alias{motherchildBMI}
\title{motherchildBMI}
\format{
A data frame with 126 rows and 2 variables:
\describe{
  \item{MotherBMI}{Variable indicating the BMI classification of the mother}
  \item{ChildBMI}{Variable indicating the BMI classification of the child}
}
}
\usage{
motherchildBMI
}
\description{
A dataset of Mothers BMI relative to Childs BMI.
}
\keyword{datasets}
