% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipleimputation.R
\name{multipleimputation}
\alias{multipleimputation}
\title{multipleimputation}
\usage{
multipleimputation(
  database,
  method = "pmm",
  imputations = 50,
  maxiterations = 0,
  restrict = TRUE,
  seed = 500
)
}
\arguments{
\item{database}{Database of variables.}

\item{method}{A string specifying the imputation method. Default is predictive mean matching but see mice help file.}

\item{imputations}{The number of imputations to perform. Default is 50.}

\item{maxiterations}{The maximum number of iterations to perform. Default is 10 x the number of iterations.}

\item{restrict}{Parameter to restrict all estimates to the range of observed values.}

\item{seed}{Seed for the random number generator to obtain reproducible results. NA will turn this parameter off.}
}
\value{
\item{database}{Database of variables with missing data replaced for the specified variables.}
}
\description{
Use multiple imputation (using the mice package) to replace missing data points. For each missing observation, multiple imputation is performed n times to create a combined multivariable modeling estimate. The database is returned with the missing data replaced with the modeling estimate.
}
\examples{

    # Run multiple imputation to replace missing cases
    # in built in dataset airquality
    data <- multipleimputation(airquality, imputations=10)

}
\author{
Matthew B. Pontifex, \email{pontifex@msu.edu}, August 18, 2017
}
