% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numpymath.R
\name{numpymath}
\alias{numpymath}
\title{numpymath}
\usage{
numpymath(vector1, operator, vector2)
}
\arguments{
\item{vector1}{data vector 1}

\item{operator}{mathmatical operator}

\item{vector2}{data vector 2}
}
\description{
Function to apply mathmatical operations, vector1 operator vector2.
}
\examples{

  outputresult <- numpymath(PlantGrowth$weight, '*', 2.2)
  
}
\author{
Matthew B. Pontifex, \email{pontifex@msu.edu}, September 13, 2021
}
