% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_correlation.R
\name{pop_correlation}
\alias{pop_correlation}
\title{pop_correlation}
\usage{
pop_correlation()
}
\description{
Popup window to compute SPSS style correlations
}
\author{
Matthew B. Pontifex, \email{pontifex@msu.edu}, May 4, 2020
}
