% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_descriptives.R
\name{pop_descriptives}
\alias{pop_descriptives}
\title{pop_descriptives}
\usage{
pop_descriptives()
}
\description{
Popup window to compute SPSS style descriptives.
}
\author{
Matthew B. Pontifex, \email{pontifex@msu.edu}, May 3, 2020
}
