% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_RmimicANOVA.R
\name{pop_RmimicANOVA}
\alias{pop_RmimicANOVA}
\title{pop_RmimicANOVA}
\usage{
pop_RmimicANOVA()
}
\description{
Popup window to compute SPSS style univariate ANOVA with effect size and confidence intervals
}
\author{
Matthew B. Pontifex, \email{pontifex@msu.edu}, May 5, 2020
}
