% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_RmimicChisquare.R
\name{pop_RmimicChisquare}
\alias{pop_RmimicChisquare}
\title{pop_RmimicChisquare}
\usage{
pop_RmimicChisquare()
}
\description{
Popup window to compute SPSS style regression analysis with effect size and confidence intervals
}
\author{
Matthew B. Pontifex, \email{pontifex@msu.edu}, May 7, 2020
}
