% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_RmimicLMcontrast.R
\name{pop_RmimicLMcontrast}
\alias{pop_RmimicLMcontrast}
\title{pop_RmimicLMcontrast}
\usage{
pop_RmimicLMcontrast()
}
\description{
Popup window to compute SPSS style regression analysis with effect size and confidence intervals
}
\author{
Matthew B. Pontifex, \email{pontifex@msu.edu}, May 5, 2020
}
