% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_RmimicMediation.R
\name{pop_RmimicMediation}
\alias{pop_RmimicMediation}
\title{pop_RmimicMediation}
\usage{
pop_RmimicMediation()
}
\description{
Popup window to compute mediation analysis with confidence intervals
}
\author{
Matthew B. Pontifex, \email{pontifex@msu.edu}, October 1, 2020
}
