% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_RmimicMLAnova.R
\name{pop_RmimicMLAnova}
\alias{pop_RmimicMLAnova}
\title{pop_RmimicMLAnova}
\usage{
pop_RmimicMLAnova()
}
\description{
Popup window to compute univariate ANOVA with effect size and confidence intervals using a multi-level model from the lme4 function
}
\author{
Matthew B. Pontifex, \email{pontifex@msu.edu}, May 11, 2020
}
