% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_ttest.R
\name{pop_ttest}
\alias{pop_ttest}
\title{pop_ttest}
\usage{
pop_ttest()
}
\description{
Popup window to compute SPSS style t-test with effect size and confidence intervals
}
\author{
Matthew B. Pontifex, \email{pontifex@msu.edu}, May 4, 2020
}
