% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_updateRmimic.R
\name{pop_updateRmimic}
\alias{pop_updateRmimic}
\title{pop_updateRmimic}
\usage{
pop_updateRmimic()
}
\description{
Popup window to update Rmimic
}
\author{
Matthew B. Pontifex, \email{pontifex@msu.edu}, Nov 2, 2021
}
