% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posthoc2text.R
\name{posthoc2text}
\alias{posthoc2text}
\title{posthoc2text}
\usage{
posthoc2text(result, studywiseAlpha = 0.05, spansize = 95,
  currentlevelout = 0)
}
\arguments{
\item{result}{Anova and posthoc test output}

\item{studywiseAlpha}{Decimal representation of alpha level. Default 0.05.}

\item{spansize}{Number of characters to include on a line.}

\item{currentlevelout}{Integer for tab indenting used in recursion.}
}
\description{
Output ANOVA posthoc results in APA style format with effect sizes and confidence intervals.
}
\author{
Matthew B. Pontifex, \email{pontifex@msu.edu}, October 30, 2019
}
