% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudobootstrapsamples.R
\name{pseudobootstrapsamples}
\alias{pseudobootstrapsamples}
\title{pseudobootstrapsamples}
\usage{
pseudobootstrapsamples(
  data,
  newsamples = 10,
  subjectid = NULL,
  method = NULL,
  seed = NA
)
}
\arguments{
\item{data}{Data frame containing the variables of interest.}

\item{newsamples}{Integer representing the number of new samples to draw.}

\item{subjectid}{Subject ID label.}

\item{method}{Parameter - Add vs Replace - to determine if new samples are added to the existing dataframe or if the returned dataframe should be fully resampled.}

\item{seed}{Seed for the random number generator to obtain reproducible results. NA will turn this parameter off.}
}
\value{
\item{database}{Database of variables with missing data replaced for the specified variables.}
}
\description{
For Exploratory Use Only. This function returns a dataframe with additional samples included drawn randomly from the existing database.
}
\examples{

    # Increase sample by 10 
    data <- pseudobootstrapsamples(data=airquality, newsamples=10, subjectid=NULL, method='Add', seed=NA)

}
\author{
Matthew B. Pontifex, \email{pontifex@msu.edu}, November 20, 2020
}
