% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RmimicAnova.R
\name{RmimicAnova}
\alias{RmimicAnova}
\title{RmimicAnova}
\usage{
RmimicAnova(
  data,
  dependentvariable = NULL,
  subjectid = NULL,
  between = NULL,
  within = NULL,
  sphericity = NULL,
  feffect = NULL,
  nonparametric = FALSE,
  posthoc = "False Discovery Rate Control",
  FDRC = 0.05,
  planned = NULL,
  suppressposthoc = NULL,
  confidenceinterval = 0.95,
  studywiseAlpha = 0.05,
  verbose = TRUE,
  verbosedescriptives = TRUE,
  posthoclimit = 6
)
}
\arguments{
\item{data}{Database containing data}

\item{dependentvariable}{Dependent Variable label}

\item{subjectid}{Subject ID label.}

\item{between}{Between subjects column labels.}

\item{within}{Within subjects column labels.}

\item{sphericity}{Parameter to select Sphericity correction method. Default is Greenhouse-Geisser. Other option is Huynh-Feldt.}

\item{feffect}{Parameter to select which eta square statistic to use for effect size computation. Default is Generalized Eta Squared. Other option is Partial Eta Squared.}

\item{nonparametric}{Parameter to select use of non parametric t tests. Default is FALSE.}

\item{posthoc}{Parameter to indicate what post-hoc comparisons should be performed. Default is False Discovery Rate Control. Other options are Bonferroni, Holm-Bonferroni, Scheffe, Sidak, Tukey, or False Discovery Rate Control.}

\item{FDRC}{Decimal representation of false discocvery rate control. Default is 0.05.}

\item{planned}{Parameter to specify an effect to show the post-hoc comparisons even if they are not significant.}

\item{suppressposthoc}{Parameter to specify a posthoc effect to ignore even if it is significant.}

\item{confidenceinterval}{Decimal representation of confidence interval. Default 0.95.}

\item{studywiseAlpha}{Decimal representation of alpha level. Default 0.05.}

\item{verbose}{Boolean operator for if interpretations of the statistics should be printed. Default is TRUE.}

\item{verbosedescriptives}{Boolean operator for if descriptive statistics should be printed. Default is TRUE.}

\item{posthoclimit}{Parameter to specify the limit for breaking down interaction terms. Default is 6 indicating a 6 way interaction would not be automatically broken down.}
}
\value{
\item{stats}{ANOVA summary table.}
\item{aov}{An aov object corresponding to the requested ANOVA.}
\item{posthocttest}{A t test summary table for posthoc decompositions.}
}
\description{
Compute SPSS style univariate ANOVA with effect size and confidence intervals using the ezANOVA function. Main effects and interactions are automatically decomposed using the specified post-hoc corrections. Subsequent posthoc ANOVAs are outputted as seperate entries in the results for clarity. Note: Data should be in long format.
}
\examples{

    # Compute ANOVA of PlantGrowth dataset
    result <- RmimicAnova(
                data = PlantGrowth,
                dependentvariable = "weight",
                subjectid = NULL,
                between = "group",
                within = NULL)

   # Compute ANOVA of ToothGrowth dataset
   result <- RmimicAnova(
                data = ToothGrowth,
                dependentvariable = "len",
                subjectid = NULL,
                between = c("dose","supp"),
                within = NULL)

}
\author{
Matthew B. Pontifex, \email{pontifex@msu.edu}, October 30, 2019
}
