% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RmimicCSSstyle.R
\name{RmimicCSSstyle}
\alias{RmimicCSSstyle}
\title{RmimicCSSstyle}
\usage{
RmimicCSSstyle()
}
\description{
Preformatted CSS style sheet
}
\author{
Matthew B. Pontifex, \email{pontifex@msu.edu}, April 14, 2025
}
