% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RmimicLMcontrast.R
\name{RmimicLMcontrast}
\alias{RmimicLMcontrast}
\title{RmimicLMcontrast}
\usage{
RmimicLMcontrast(
  fit,
  altfit,
  confidenceinterval = 0.95,
  studywiseAlpha = 0.05,
  verbose = TRUE
)
}
\arguments{
\item{fit}{fit from a basic linear model}

\item{altfit}{fit from a linear model of interest}

\item{confidenceinterval}{Decimal representation of confidence interval. Default 0.95.}

\item{studywiseAlpha}{Decimal representation of alpha level. Default 0.05.}

\item{verbose}{Boolean operator for if interpretations of the statistics should be printed. Default is TRUE.}
}
\value{
\item{stats}{ANOVA summary table.}
\item{changestats}{ANOVA summary table of change statistics.}
\item{coefficients}{A summary table of regression coefficients.}
}
\description{
Compute SPSS style results for regression analysis with effect size and confidence intervals. This function takes stats::lm fits for a base model and the model of interest and calculates statistics for the model of interest relative to the base model.
}
\examples{

    # Compute Model from mtcars dataset
    basefit <- lm(mpg ~ am + wt, data = mtcars)
    fit <- lm(mpg ~ am + wt + qsec, data = mtcars)
    regresult <- RmimicLMcontrast(basefit, fit, confidenceinterval=0.95, studywiseAlpha=0.05, verbose=TRUE)

}
\author{
Matthew B. Pontifex, \email{pontifex@msu.edu}, January 31, 2020
}
