% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rmimiclsmeans.R
\name{Rmimiclsmeans}
\alias{Rmimiclsmeans}
\title{Rmimiclsmeans}
\usage{
Rmimiclsmeans(
  fit,
  data,
  dependentvariable = NULL,
  subjectid = NULL,
  between = NULL,
  within = NULL,
  df = NULL,
  posthoc = NULL,
  criticaldiff = NULL,
  confidenceinterval = 0.95,
  studywiseAlpha = 0.05,
  verbose = TRUE
)
}
\arguments{
\item{fit}{fit from a lmer model.}

\item{data}{Data frame in long format containing the dependent variable in one column and subject id, between, and within subject variables in other columns.}

\item{dependentvariable}{Variable name or list of variables to use as dependent variables.}

\item{subjectid}{Variable name corresponding to the subject id column in the data. Required for paired tests of within subjects variables, encouraged otherwise.}

\item{between}{Variable name or list of variables to use to compute independent samples t tests.}

\item{within}{Variable name or list of variables to use to compute paired samples t tests.}

\item{df}{Parameter to indicate what degrees of freedom approximation should be used. Default is Kenward-Roger. Other option is Shattertwaite.}

\item{posthoc}{Parameter to determine what post-hoc comparison approach to use. Options are Bonferroni, Sidak, or Holm-Bonferroni.}

\item{criticaldiff}{Parameter to specify the critical difference used in Tukey and Scheffe post-hoc comparison approaches.}

\item{confidenceinterval}{Parameter to control the confidence interval. Default is 0.95.}

\item{studywiseAlpha}{Parameter to control the study wise alpha for use in the post hoc compairsons. Default is 0.05.}

\item{verbose}{Parameter to print all output to console. Default is TRUE.}
}
\value{
\item{descriptives}{Data table of descriptive statistics.}
\item{stats}{Data table of t test statistics.}
}
\description{
Subprocess to RmimicMLAnova function to extract t-test with effect size and confidence intervals.
}
\author{
Matthew B. Pontifex, \email{pontifex@msu.edu}, May 10, 2020
}
