% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RmimicMLAnova.R
\name{RmimicMLAnova}
\alias{RmimicMLAnova}
\title{RmimicMLAnova}
\usage{
RmimicMLAnova(
  data,
  dependentvariable = NULL,
  subjectid = NULL,
  between = NULL,
  within = NULL,
  covariates = NULL,
  randomintercept = NULL,
  randomslope = NULL,
  df = NULL,
  posthoc = "False Discovery Rate Control",
  FDRC = 0.05,
  planned = NULL,
  suppressposthoc = NULL,
  confidenceinterval = 0.95,
  studywiseAlpha = 0.05,
  verbose = TRUE,
  verbosedescriptives = TRUE,
  posthoclimit = 6
)
}
\arguments{
\item{data}{Database containing data}

\item{dependentvariable}{Dependent Variable label}

\item{subjectid}{Subject ID label.}

\item{between}{Between subjects column labels.}

\item{within}{Within subjects column labels.}

\item{covariates}{Non interactive variables to be included in the analysis as covariates.}

\item{randomintercept}{Parameter to indicate random intercepts.}

\item{randomslope}{Parameter to indicate random slopes.}

\item{df}{Parameter to indicate what degrees of freedom approximation should be used. Default is Kenward-Roger. Other option is Shattertwaite.}

\item{posthoc}{Parameter to indicate what post-hoc comparisons should be performed. Default is False Discovery Rate Control. Other options are Bonferroni, Holm-Bonferroni, Scheffe, Sidak, Tukey, or False Discovery Rate Control.}

\item{FDRC}{Decimal representation of false discocvery rate control. Default is 0.05.}

\item{planned}{Parameter to specify an effect to show the post-hoc comparisons even if they are not significant.}

\item{suppressposthoc}{Parameter to specify a posthoc effect to ignore even if it is significant.}

\item{confidenceinterval}{Decimal representation of confidence interval. Default 0.95.}

\item{studywiseAlpha}{Decimal representation of alpha level. Default 0.05.}

\item{verbose}{Boolean operator for if interpretations of the statistics should be printed. Default is TRUE.}

\item{verbosedescriptives}{Boolean operator for if descriptive statistics should be printed. Default is TRUE.}

\item{posthoclimit}{Parameter to specify the limit for breaking down interaction terms. Default is 6 indicating a 6 way interaction would not be automatically broken down.}
}
\value{
\item{stats}{ANOVA summary table.}
\item{aov}{An aov object corresponding to the requested ANOVA.}
\item{posthocttest}{A t test summary table for posthoc decompositions.}
}
\description{
Compute SPSS style univariate ANOVA with effect size and confidence intervals using a multi-level model from the lme4 function. Main effects and interactions are automatically decomposed using the specified post-hoc corrections. Note: Data should be in long format. As this function will conduct full post-hoc decompositions, the more factors assess the longer processing will take.
}
\author{
Matthew B. Pontifex, \email{pontifex@msu.edu}, May 11, 2020
}
