% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RmimicTtest.R
\name{RmimicTtest}
\alias{RmimicTtest}
\title{RmimicTtest}
\usage{
RmimicTtest(data, dependentvariable = NULL, subjectid = NULL,
  between = NULL, within = NULL, nonparametric = FALSE,
  collapse = TRUE, posthoc = NULL, confidenceinterval = 0.95,
  studywiseAlpha = 0.05, verbose = TRUE)
}
\arguments{
\item{data}{Data frame in long format containing the dependent variable in one column and subject id, between, and within subject variables in other columns.}

\item{dependentvariable}{Variable name or list of variables to use as dependent variables.}

\item{subjectid}{Variable name corresponding to the subject id column in the data.}

\item{between}{Variable name or list of variables to use to compute independent samples t tests.}

\item{within}{Variable name or list of variables to use to compute paired samples t tests.}

\item{nonparametric}{Parameter to determine if non parametric tests should be run. Default is FALSE.}

\item{collapse}{Paramater to determine if multiple observations should be collapsed to a single observation for each participant. Default is TRUE.}

\item{posthoc}{Parameter to determine what post-hoc comparison approach to use. Options are Bonferroni, Sidak, or Holm-Bonferroni.}

\item{confidenceinterval}{Parameter to control the confidence interval. Default is 0.95.}

\item{studywiseAlpha}{Parameter to control the study wise alpha for use in the post hoc compairsons. Default is 0.05.}

\item{verbose}{Parameter to print all output to console. Default is TRUE.}
}
\value{
\item{descriptives}{Data table of descriptive statistics.}
\item{stats}{Data table of t test statistics.}
}
\description{
Compute SPSS style t-test with effect size and confidence intervals. Optional parameters are also provided to compute non-parametric t-tests with appropriate non-parametric effect size estimates. For parametric test the function automatically determines if the variances are equal using levene's test and outputs the correct statistcs.
}
\examples{

  ttestresult <- RmimicTtest(PlantGrowth, 
          dependentvariable='weight', 
          subjectid=NULL, 
          between='group', 
          within=NULL, 
          nonparametric=FALSE,
          posthoc="Holm-Bonferroni")

}
\author{
Matthew B. Pontifex, \email{pontifex@msu.edu}, October 13, 2019
}
