% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subprocessRmimicChisquare.R
\name{subprocessRmimicChisquare}
\alias{subprocessRmimicChisquare}
\title{subprocessRmimicChisquare}
\usage{
subprocessRmimicChisquare(variables = FALSE, data = FALSE)
}
\arguments{
\item{variables}{Variable name or list of variables to compute descriptives for.}

\item{data}{Data frame containing the variables of interest.}
}
\value{
\item{stats}{Summary of analysis.}
\item{tabularoutput}{Chi square table.}
}
\description{
Subprocess to compute Chi-square analysis. For samples less than 1000, Fishers exact test statistic is used.
}
\author{
Matthew B. Pontifex, \email{pontifex@msu.edu}, March 6, 2020
}
