% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summateacrosscolumns.R
\name{summateacrosscolumns}
\alias{summateacrosscolumns}
\title{summateacrosscolumns}
\usage{
summateacrosscolumns(workbook, columnnames, makenumeric = TRUE, count = FALSE)
}
\arguments{
\item{workbook}{data frame that contains data}

\item{columnnames}{column names to sum across}

\item{makenumeric}{boolean parameter to force values to numeric}

\item{count}{boolean parameter to return the number of non empty values}
}
\description{
Function to summate across columns
}
\author{
Matthew B. Pontifex, \email{pontifex@msu.edu}, September 13, 2021
}
