% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table2frame.R
\name{table2frame}
\alias{table2frame}
\title{table2frame}
\usage{
table2frame(data)
}
\arguments{
\item{data}{Table with a column labeled freq.}
}
\description{
Helper function to blow up a table into a data frame.
}
\author{
Matthew B. Pontifex, \email{pontifex@msu.edu}, May 19, 2020
}
