% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest2text.R
\name{ttest2text}
\alias{ttest2text}
\title{ttest2text}
\usage{
ttest2text(datalist, verbose = TRUE)
}
\arguments{
\item{datalist}{t-test list output}

\item{verbose}{Parameter to print all output to console. Default is TRUE.}
}
\value{
A list with the following elements:
\item{statistics}{Data table of statistics.}
\item{text}{Statistics formatted in text.}
}
\description{
Output t-test results in APA style format.
}
\author{
Matthew B. Pontifex, \email{pontifex@msu.edu}, October 7, 2019
}
