% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttesttopval.R
\name{ttesttopval}
\alias{ttesttopval}
\title{ttesttopval}
\usage{
ttesttopval(tval = FALSE, df = FALSE, tail = NULL)
}
\arguments{
\item{tval}{T test value.}

\item{df}{Degrees of Freedom.}

\item{tail}{Tail, default is two tailed. Options for Right or Left.}
}
\value{
\item{stats}{Summary of analysis.}
}
\description{
P value for a t statistic.
}
\author{
Matthew B. Pontifex, \email{pontifex@msu.edu}, April 7, 2022
}
