% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typewriter.R
\name{typewriter}
\alias{typewriter}
\title{typewriter}
\usage{
typewriter(textstring, tabs, spaces, characters, indent)
}
\arguments{
\item{textstring}{String to print to console.}

\item{tabs}{How many tabs should be printed before the text.}

\item{spaces}{Following tab indents, how many spaces should be printed before the text.}

\item{characters}{How many characters wide should the output be. Default is options()$width .}

\item{indent}{Parameter to control the indent. Positive numbers indicate an indent. Negative numbers indicate hanging format.}
}
\description{
Function to control the text outputted to the console to create a consistent indent.
}
\examples{
    # Example text block
    textblock <- "Lorem ipsum dolor sit amet, consectetur adipiscing elit,
                  sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.
                  Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris
                  nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in
                  reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla
                  pariatur. Excepteur sint occaecat cupidatat non proident, sunt in
                  culpa qui officia deserunt mollit anim id est laborum."

    # Output indented text 3 tabs in
    typewriter(textblock, tabs=3, spaces=0, characters=100, indent=TRUE)

    # Output hanging text 1 tabs in with 70 character width
    typewriter(textblock, tabs=1, spaces=0, characters=70, indent='hanging')

    # Output indented text 2 tabs and 3 spaces in, with 100 character width
    typewriter(textblock, tabs=2, spaces=3, characters=100, indent=FALSE)


}
\author{
Matthew B. Pontifex, \email{pontifex@msu.edu}, August 31, 2017
}
