% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzyP.R
\name{fuzzyP}
\alias{fuzzyP}
\title{fuzzyP}
\usage{
fuzzyP(datain, studywiseAlpha = 0.05, html = FALSE)
}
\arguments{
\item{datain}{Significance value.}

\item{html}{Boolean parameter for output in html format.}

\item{alpha}{Critical value for alpha.}
}
\value{
A list with the following elements:
\item{interpret}{Rounded p value for use in interpretations.}
\item{report}{Rounded p value for use in summary reports.}
\item{exact}{Exact p value.}
\item{modifier}{Equals sign unless the p value is less than .001.}
\item{significance}{Boolean for if the value is less than or equal to the critical alpha.}
}
\description{
Function to properly round p values. Returns several values allowing \cr the user to choose what rounded value is appropriate.
}
\examples{
    # Round a randomly generated number between 0 and 1.
    outPvalue <- fuzzyP(stats::runif(1))

}
\author{
Matthew B. Pontifex, \email{pontifex@msu.edu}, July 28, 2017
}
