% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posthoc2text.R
\name{posthoc2text}
\alias{posthoc2text}
\title{posthoc2text}
\usage{
posthoc2text(
  result,
  studywiseAlpha = 0.05,
  spansize = 95,
  currentlevelout = 0,
  posthoclimit = 6,
  planned = NULL,
  suppressposthoc = NULL
)
}
\arguments{
\item{result}{Anova and posthoc test output}

\item{studywiseAlpha}{Decimal representation of alpha level. Default 0.05.}

\item{spansize}{Number of characters to include on a line.}

\item{currentlevelout}{Integer for tab indenting used in recursion.}

\item{posthoclimit}{Parameter to specify the limit for breaking down interaction terms. Default is 6 indicating a 6 way interaction would not be automatically broken down.}

\item{planned}{Parameter to specify an effect to show the post-hoc comparisons even if they are not significant.}

\item{suppressposthoc}{Parameter to specify a posthoc effect to ignore even if it is significant.}
}
\description{
Output ANOVA posthoc results in APA style format with effect sizes and confidence intervals.
}
\author{
Matthew B. Pontifex, \email{pontifex@msu.edu}, October 30, 2019
}
